move logging back to util
