+++ torchdynamo/utils.py
@@ -12,6 +12,7 @@ import logging
 import logging.config
 import math
 import operator
+import os
 import pstats
 import re
 import sys
@@ -115,8 +116,60 @@ def compile_times(repr="str", aggregate=False):
         return headers, values
 
 
+# Return all loggers that torchdynamo is responsible for
+def get_loggers():
+    return [
+        logging.getLogger("torchdynamo"),
+        logging.getLogger("torchinductor"),
+    ]
+
+
+# Set the level of all loggers that torchdynamo is responsible for
+def set_loggers_level(level):
+    for logger in get_loggers():
+        logger.setLevel(level)
+
+
+LOGGING_CONFIG = {
+    "version": 1,
+    "formatters": {
+        "torchdynamo_format": {"format": "%(name)s: [%(levelname)s] %(message)s"},
+    },
+    "handlers": {
+        "torchdynamo_console": {
+            "class": "logging.StreamHandler",
+            "level": "DEBUG",
+            "formatter": "torchdynamo_format",
+            "stream": "ext://sys.stdout",
+        },
+    },
+    "loggers": {
+        "torchdynamo": {
+            "level": "DEBUG",
+            "handlers": ["torchdynamo_console"],
+            "propagate": False,
+        },
+        "torchinductor": {
+            "level": "DEBUG",
+            "handlers": ["torchdynamo_console"],
+            "propagate": False,
+        },
+    },
+    "disable_existing_loggers": False,
+}
+
+
+# initialize torchdynamo loggers
 def init_logging():
-    torchdynamo.logging.init_logging(config.log_level, file_name=config.log_file_name)
+    if "PYTEST_CURRENT_TEST" not in os.environ:
+        logging.config.dictConfig(LOGGING_CONFIG)
+        if config.log_file_name is not None:
+            log_file = logging.FileHandler(config.log_file_name)
+            log_file.setLevel(config.log_level)
+            for logger in get_loggers():
+                logger.addHandler(log_file)
+
+    set_loggers_level(config.log_level)
 
 
 # filter out all frames after entering dynamo
